app/faxplain/requirements.txt
flask==1.1.2 
gunicorn==20.1.0
torch==1.8.1
gensim==4.0.1
transformers==3.0.2
pyyaml==5.4.1
